Able to GET users w/ specific ID

+++ migrations/20230718173657_create_user_table.js
@@ -4,7 +4,7 @@
  */
 exports.up = function(knex) {
   return knex.schema.createTable('users', table => {
-    table.increments('id').primary();
+    table.BigInt('id').primary();
     table.string('first_name');
     table.string('last_name');
     table.string('username');
